+++ packages/orq-rc/src/funcs/apiKeysGet.ts
@@ -3,7 +3,7 @@
  */
 
 import { OrqCore } from "../core.js";
-import { encodeSimple } from "../lib/encodings.js";
+import { encodeFormQuery, encodeSimple } from "../lib/encodings.js";
 import { matchStatusCode } from "../lib/http.js";
 import * as M from "../lib/matchers.js";
 import { compactMap } from "../lib/primitives.js";
@@ -95,6 +95,10 @@ async function $do(
   };
   const path = pathToFunc("/v2/api-keys/{api_key_id}")(pathParams);
 
+  const query = encodeFormQuery({
+    "include_budget": payload.include_budget,
+  });
+
   const headers = new Headers(compactMap({
     Accept: "application/json",
   }));
@@ -124,6 +128,7 @@ async function $do(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
+    query: query,
     body: body,
     userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || 600000,