+++ packages/orq-rc/src/funcs/identitiesList.ts
@@ -93,6 +93,7 @@ async function $do(
   const query = encodeFormQuery({
     "ending_before": payload?.ending_before,
     "filter_by.tags": payload?.["filter_by.tags"],
+    "include_budget": payload?.include_budget,
     "include_metrics": payload?.include_metrics,
     "limit": payload?.limit,
     "search": payload?.search,