+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3975,7 +3975,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KWA3J1DPQ1DTWSY9ZYHSXY1R"),
+  id: z.string().default("01KWB8CG08NV635TYYZBP5V813"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),