+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-29T16:31:22.968Z",
+    "2026-06-30T03:15:02.428Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {